requirements.txt
appnope==0.1.0
bleach==1.5.0
boto3==1.4.8
botocore==1.8.8
cycler==0.10.0
decorator==4.1.2
dj-database-url==0.4.2
Django==1.11.6
django-storages==1.6.5
docutils==0.14
entrypoints==0.2.3
gunicorn==19.7.1
html5lib==0.9999999
ipykernel==4.6.1
ipython==6.2.1
ipython-genutils==0.2.0
ipywidgets==7.0.1
jedi==0.11.0
Jinja2==2.9.6
jmespath==0.9.3
jsonschema==2.6.0
jupyter==1.0.0
jupyter-client==5.1.0
jupyter-console==5.2.0
jupyter-core==4.3.0
Markdown==2.6.9
MarkupSafe==1.0
matplotlib==2.1.0
mistune==0.7.4
nbconvert==5.3.1
nbformat==4.4.0
notebook==5.1.0
numpy==1.13.3
olefile==0.44
pandas==0.20.3
pandocfilters==1.4.2
parso==0.1.0
pexpect==4.2.1
pickleshare==0.7.4
Pillow==4.3.0
prompt-toolkit==1.0.15
protobuf==3.4.0
psycopg2==2.7.3.2
ptyprocess==0.5.2
Pygments==2.2.0
pyparsing==2.2.0
python-dateutil==2.6.1
python-decouple==3.1
pytz==2017.2
pyzmq==16.0.2
qtconsole==4.3.1
s3transfer==0.1.12
scikit-learn==0.19.0
scipy==0.19.1
simplegeneric==0.8.1
six==1.11.0
sklearn==0.0
tensorflow==1.3.0
tensorflow-tensorboard==0.1.8
terminado==0.6
testpath==0.3.1
tornado==4.5.2
traitlets==4.3.2
virtualenv==15.1.0
wcwidth==0.1.7
Werkzeug==0.12.2
whitenoise==3.3.1
widgetsnbextension==3.0.3
